endpoin update

+++ controllers/users.controller.js
@@ -40,10 +40,10 @@ const createUser = async (req, res) => {
   });
 
   if (data) {
-    const response = { status: 200, message: "success create user" };
+    const response = { status: 200, message: "success to create user" };
     return res.status(200).json(response);
   } else {
-    const response = { status: 400, message: "unsuccess create user" };
+    const response = { status: 400, message: "failed to create user" };
     return res.status(400).json(response);
   }
 };
@@ -58,10 +58,74 @@ const getAllUser = async (req, res) => {
 };
 
 const deleteUser = async (req, res) => {
+  const userIsReady = await usersModel.findOne({ where: { id: req.params.id } });
+
+  if (!userIsReady) {
+    const response = { status: 400, message: "user data not found!" };
+    return res.status(400).json(response);
+  }
+
   await usersModel.destroy({ where: { id: req.params.id } });
 
   const response = { status: 201, data: "data deleted" };
   return res.status(201).json(response);
 };
 
-module.exports = { createUser, getAllUser, deleteUser };
+const updateUser = async (req, res) => {
+  const userIsReady = await usersModel.findOne({ where: { id: req.params.id } });
+
+  if (!userIsReady) {
+    const response = { status: 400, message: "user data not found!" };
+    return res.status(400).json(response);
+  }
+
+  const schema = {
+    fullName: "string|empty:false",
+    email: "email",
+    password: "string|min:8",
+  };
+
+  const validate = v.validate(req.body, schema);
+
+  if (validate.length) {
+    const response = {
+      status: 400,
+      errors: validate,
+    };
+    return res.status(400).json(response);
+  }
+
+  const { fullName, email, password, userStatus, userImg } = req.body;
+
+  const emailAlreadyExist = await usersModel.findOne({ where: { email } });
+
+  if (emailAlreadyExist && emailAlreadyExist.id != req.params.id) {
+    const response = { status: 409, message: "email already exist!" };
+    return res.status(409).json(response);
+  }
+
+  const encryptPassword = await bcrypt.hash(password, 10);
+
+  const data = await usersModel.update(
+    {
+      fullName,
+      email,
+      password: encryptPassword,
+      userStatus,
+      userImg,
+    },
+    {
+      where: { id: req.params.id },
+    }
+  );
+
+  if (data) {
+    const response = { status: 200, message: "success to update data user" };
+    return res.status(200).json(response);
+  } else {
+    const response = { status: 400, message: "failed to update user" };
+    return res.status(400).json(response);
+  }
+};
+
+module.exports = { createUser, getAllUser, deleteUser, updateUser };
